Update server.js
add helper function to render content conditionally. This function is used to render rows on the /allProjects page depending on which statusFilter is being applied.

Also, update the format that date is being entered into DB when user/faculty enters new record. This allows for comparison operations to be performed (like filtering for specific dates).

Additionally, change request from POST to GET when search form is submitted. When staying on the search page for long enough the server would automatically reload the page by submitting a GET request. However, one did not exist for /faculty/search, leading to a 404 error. Converting the post request to get fixed this error while maintaining existing functionality.